Default ONSET_SEC and OFFSET_SEC values to music21 symbolic timepoints
This means that the default is not in actual seconds despite the column names.
ONSET_SEC is the same as ONSET and OFFSET_SEC is equal to ONSET + DURATION. The
idea is that these will get filled in later when actual second timings are
brought in from a json file.